Apply comments from review for https://github.com/plone/plone.formwidget.namedfile/pull/43
